New Gui Designs / New Syncing
- Showing unloaded connections in the gui
- Options for viewing the current connections
- Displays the input/outputs of the flux plug in the gui (need to do it
for flux point still)
- Fixed Network Statistics

+++ src/main/java/sonar/flux/connection/transfer/handlers/FluxTransferHandler.java
@@ -13,12 +13,12 @@ public abstract class FluxTransferHandler<T extends IFlux> extends BaseTransferH
 	}
 
 	@Override
-	public double getMaxRemove() {
+	public long getMaxRemove() {
 		return flux.getTransferLimit();
 	}
 
 	@Override
-	public double getMaxAdd() {
+	public long getMaxAdd() {
 		return flux.getTransferLimit();
 	}
 }
